Move micros rebase functions to RebaseDateTime
