+++ sql/core/src/main/scala/org/apache/spark/sql/execution/datasources/parquet/ParquetRowConverter.scala
@@ -35,7 +35,7 @@ import org.apache.spark.sql.catalyst.InternalRow
 import org.apache.spark.sql.catalyst.expressions._
 import org.apache.spark.sql.catalyst.util.{ArrayBasedMapData, CaseInsensitiveMap, DateTimeUtils, GenericArrayData}
 import org.apache.spark.sql.catalyst.util.DateTimeUtils.SQLTimestamp
-import org.apache.spark.sql.catalyst.util.RebaseDateTime.rebaseJulianToGregorianDays
+import org.apache.spark.sql.catalyst.util.RebaseDateTime._
 import org.apache.spark.sql.internal.SQLConf
 import org.apache.spark.sql.types._
 import org.apache.spark.unsafe.types.UTF8String
@@ -278,7 +278,7 @@ private[parquet] class ParquetRowConverter(
         if (rebaseDateTime) {
           new ParquetPrimitiveConverter(updater) {
             override def addLong(value: Long): Unit = {
-              val rebased = DateTimeUtils.rebaseJulianToGregorianMicros(value)
+              val rebased = rebaseJulianToGregorianMicros(value)
               updater.setLong(rebased)
             }
           }
@@ -295,7 +295,7 @@ private[parquet] class ParquetRowConverter(
           new ParquetPrimitiveConverter(updater) {
             override def addLong(value: Long): Unit = {
               val micros = DateTimeUtils.millisToMicros(value)
-              val rebased = DateTimeUtils.rebaseJulianToGregorianMicros(micros)
+              val rebased = rebaseJulianToGregorianMicros(micros)
               updater.setLong(rebased)
             }
           }
