+++ sql/core/src/main/scala/org/apache/spark/sql/execution/datasources/parquet/ParquetWriteSupport.scala
@@ -35,7 +35,7 @@ import org.apache.spark.sql.SPARK_VERSION_METADATA_KEY
 import org.apache.spark.sql.catalyst.InternalRow
 import org.apache.spark.sql.catalyst.expressions.SpecializedGetters
 import org.apache.spark.sql.catalyst.util.DateTimeUtils
-import org.apache.spark.sql.catalyst.util.RebaseDateTime.rebaseGregorianToJulianDays
+import org.apache.spark.sql.catalyst.util.RebaseDateTime._
 import org.apache.spark.sql.internal.SQLConf
 import org.apache.spark.sql.types._
 
@@ -189,7 +189,7 @@ class ParquetWriteSupport extends WriteSupport[InternalRow] with Logging {
 
           case SQLConf.ParquetOutputTimestampType.TIMESTAMP_MICROS if rebaseDateTime =>
             (row: SpecializedGetters, ordinal: Int) =>
-              val rebasedMicros = DateTimeUtils.rebaseGregorianToJulianMicros(row.getLong(ordinal))
+              val rebasedMicros = rebaseGregorianToJulianMicros(row.getLong(ordinal))
               recordConsumer.addLong(rebasedMicros)
 
           case SQLConf.ParquetOutputTimestampType.TIMESTAMP_MICROS =>
@@ -198,7 +198,7 @@ class ParquetWriteSupport extends WriteSupport[InternalRow] with Logging {
 
           case SQLConf.ParquetOutputTimestampType.TIMESTAMP_MILLIS if rebaseDateTime =>
             (row: SpecializedGetters, ordinal: Int) =>
-              val rebasedMicros = DateTimeUtils.rebaseGregorianToJulianMicros(row.getLong(ordinal))
+              val rebasedMicros = rebaseGregorianToJulianMicros(row.getLong(ordinal))
               val millis = DateTimeUtils.microsToMillis(rebasedMicros)
               recordConsumer.addLong(millis)
